refactor(finance): extract remaining pure tax-remittance helpers into the domain-core module (BI-OPT-FAT-ACTIONS — tax-remittance remainder)
Pure behavior-preserving MOVE of the remaining side-effect-free logic out of
apps/web/lib/actions/tax-remittance.ts into the existing domain-core module
apps/web/lib/finance/tax-remittance-core.ts, completing the extraction begun in
Slice A and mirroring the EA (#3118), CRM (#3125), Build (#3126), agent-coworker
(#3128) and platform-dev-config (#3129) slices.

Relocated verbatim (bodies unchanged) into tax-remittance-core.ts:
  - Structural record shapes: TaxRegistrationRecord, LiabilityDraft,
    ManagedTaxIssueDraft (all inline structural types, not prisma-derived).
  - Liability-draft builders: buildInvoiceLiabilityDrafts, buildBillLiabilityDrafts.
  - Period/date math: computeNextCronRun, periodMonthsForFrequency.
  - Projections: buildFilingPacketNotes, buildManagedTaxIssues, buildCoworkerGuide.

Orchestration (auth, prisma, revalidatePath, notifications, the "use server"
wrappers, reconcileTaxIssues/loadTaxWorkspaceState/persistLiabilityDrafts which
touch the DB) STAYS in the action file and imports the helpers back. The managed-
issue/coworker-guide projections read a structural profile shape; the action passes
its full OrganizationTaxProfile record (a structural supertype) unchanged.

taxableBaseFromLine was only used by the two moved builders, so it is no longer
imported back into the action file — zero unused imports. No type re-export from
the "use server" module (moved types are imported directly from core).

Action file: 1788 -> 1381 LOC (ratcheted down in scripts/module-size-baseline.txt).
Core module: 116 -> 551 LOC (< 800). Added focused unit tests for the extracted
helpers (18 assertions); the existing behavior-pin action test (19 tests) passes
unchanged. Behavior byte-for-byte preserved.

Process-Spine-Decision: pure source-only helper relocation (no schema, migration, or contract change); behavior byte-for-byte preserved and pinned by the unchanged existing action tests — no new spec/plan/doc required, mirroring the prior BI-OPT-FAT-ACTIONS slices.
Design-Grounding-Decision: BI-OPT-FAT-ACTIONS
