+++ scripts/module-size-baseline.txt
@@ -29,7 +29,7 @@ apps/web/lib/actions/platform-dev-config.ts	1331
 apps/web/lib/actions/promotions.self-upgrade.test.ts	1022
 apps/web/lib/actions/promotions.ts	948
 apps/web/lib/actions/tax-remittance.test.ts	1080
-apps/web/lib/actions/tax-remittance.ts	1788
+apps/web/lib/actions/tax-remittance.ts	1381
 apps/web/lib/ai-operations-map/load-map-data.test.ts	943
 apps/web/lib/deliberation/orchestrator.ts	808
 apps/web/lib/explore/build-process-matrix.ts	811
